// generate-data-script/src/shared/types.ts
export type Wrapped2022 = Record<
  number,
  {
    emojis_view: { emojis: any[] };
    topics_view: { entities: string[]; counts: number[] };
    topics_positive_view: { entities: string[]; counts: number[] };
    responded_by_view: { top_replies: any[] };
    top_domains_view: { top_domains: Array<{ domain: string; count: number }> };
    first_cast_view: { cast: FCCast | null };
    top_casts_view: { top_casts: any[] };
    // num_deleted_casts_view: { count: number };
    profile_fans_view: { fans: any[] };
    casted_images_view: { data: Array<{ url: string; hash: string }> };
    num_likes_view: { count: number };
    casts_over_time_view: {
      data: Array<{ week: string; count: number }>;
      max: number;
    };
    num_castaways_view: { count: number };
    persona_percentage?: string;
    persona:
      | "questioncaster"
      | "lurkcaster"
      | "ghostcaster"
      | "gramcaster"
      | "replyguy"
      | "shillcaster"
      | "threadcastor"
      | "botcaster"
      | "castaway"
      | "cryptocaster"
      | "birdappcaster";
    personas_view: {
      // [decimal percentage score, base count of items matching]
      questioncaster: [number, number];
      lurkcaster: [number, number];
      ghostcaster: [number, number];
      gramcaster: [number, number];
      replyguy: [number, number];
      shillcaster: [number, number];
      threadcastor: [number, number];
      botcaster: [number, number];
      castaway: [number, number];
      cryptocaster: [number, number];
      birdappcaster: [number, number];
    };
    num_threads_view: { count: number };
    num_questions_view: { count: number };
    num_replies_view: { count: number };
    num_gms_view: { count: number };
    num_liked_by_view: { count: number };
    num_casts_view: { count: number };
    responded_to_view: { top_replies: any[] };
    user_card_view: Pick<
      FCProfile,
      "avatar_url" | "display_name" | "username" | "fid" | "bio"
    >;
    user_number_view: { number: number | string };
  }
>;
// ...
export interface FlattenedProfile {
  // node-postgres serializes int8 to string as it's possibly bigger than Number.MAX_SAFE_NUMBER
  fid: number | string;
  address: string | null;
  username?: string | null;
  display_name?: string | null;
  avatar_url?: string | null;
  referrer?: string | null;
  avatar_verified?: boolean;
  followers?: number;
  following?: number;
  bio?: string | null;
  registered_at?: Date | string;
  updated_at?: Date | string;
}

export interface FlattenedCast {
  type: "text-short";
  published_at: Date;
  /** only tracks a full reindex, that is from discove/farcaster-indexer-v2/src/tasks/index-casts.ts **/
  last_reindexed_at: Date;
  recasted_cast_hash: string | null;
  is_recast: boolean;
  deleted: boolean;

  username: string | null;
  // node-postgres serializes int8 to string as it's possibly bigger than Number.MAX_SAFE_NUMBER
  author_fid: number | string;
  text: string;

  reply_parent_hash: string | null;
  reply_parent_username: string | null;
  // node-postgres serializes int8 to string as it's possibly bigger than Number.MAX_SAFE_NUMBER
  reply_parent_fid: number | null | string;

  hash: string;
  thread_hash: string;
  display_name: string | null;
  avatar_url: string | null;
  avatar_verified: boolean;
  mentions: JSON | any;

  replies: number | null;
  reactions: number | null;
  recasts: number | null;
  watches: number | null;

  recasters: JSON | any;
}

export type CustomCastMetrics = {
  new: number;
  top: number;
  casterengagement: number;
  toplikes: number;
  toprecasts: number;
  topwatches: number;
  topreactions: number;
  topreplies: number;
  topfollowers: number;
  topfollowernormalized: number;
  hot: number;
  hotfollowernormalized: number;
  hotfrequencynormalized: number;
};

// export interface EmailSubscription {
//   email: string;
//   subscriptions: {
//     weekly_cove_updates?: Record<string, boolean>;
//   };
//   is_suppressed: boolean;
//   created_at: string;
//   updated_at: string;
//   management_secret_token: string;
// }

/** this is a subset of the properties */
export interface OpenGraphType {
  url?: string;
  type?: string;
  image?: string;
  title?: string;
  locale?: string;
  favicon?: string;
  imageType?: string;
  site_name?: string;
  imageWidth?: string;
  description?: string;
  imageHeight?: string;
  articleModifiedTime?: string;
  video?: string;
  videoType?: string;
  videoWidth?: string;
  videoHeight?: string;
}

export interface FCCast extends FlattenedCast {
  reply_data?: FCCast[];
  og: {
    last_updated?: string;
    urls?: Array<{ url: string; last_status: number; tags: OpenGraphType }>;
  };
  reply_to_data?: FCCast;
  recast_data?: FCCast;
  weighted_keywords?: unknown;
  custom_metrics?: {
    custom_profile_metrics?: FCProfile["custom_metrics"];
    custom_cast_metrics?: CustomCastMetrics;
  };
}

export interface FCProfile extends FlattenedProfile {
  weighted_keywords?: unknown;
  // Datestring

  custom_metrics?: {
    // datestring
    last_known_active?: Date | string;
    total_casts: number;

    // registered_at
    new: number;
    total_recasts: number;
    recasts_score: number;
    // hard to track without tracking followers
    hot: number;
    hot_by_engagement: number;
    total_likes: number;
    // top?
    num_followers: number;
    num_following: number;
    frequency_score: number;
    engagement_score: number;
    per_follower_engagement_score: number;
  };
}